requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
docker==7.0.0
boto3==1.34.0
paramiko==3.4.0
apscheduler==3.10.4
sqlalchemy==2.0.29
aiosqlite==0.20.0
pydantic-settings==2.2.1
python-multipart==0.0.9
aiofiles==23.2.1
aiosmtplib==3.0.1
